+++ src/cspark/cli/_main.py
@@ -1,12 +1,15 @@
 import click
-import cspark.sdk
+from cspark.sdk import sdk_version
 
 from ._commands import register_commands
+
+__version__ = '0.1.0-beta'
+__title__ = f'Coherent Spark CLI {__version__} (sdk v{sdk_version})'
 
 
 @click.group(invoke_without_command=True)
 @click.pass_context
-@click.version_option(message=f'Coherent Spark CLI v0.1.0 (sdk v{cspark.sdk.sdk_version})')
+@click.version_option(__version__, '-v', '--version', message=__title__)
 def main(ctx: click.Context) -> None:
     if ctx.invoked_subcommand is None:
         click.echo('Interact with Coherent Spark APIs from the command line.\n')